fix the situation where the gw launcher is in the gw.exe directory

+++ GW Launcher/MulticlientPatch.cs
@@ -94,7 +94,7 @@ internal static class MulticlientPatch
                 if (File.Exists(gmoddll_modlist_path))
 					gmoddll_modlist_original_contents = File.ReadAllText(gmoddll_modlist_path);
 				File.WriteAllText(gmoddll_modlist_path, texmods);
-				if (File.Exists(gwexe_modlist_path))
+				if (File.Exists(gwexe_modlist_path) && gmoddll_modlist_path != gwexe_modlist_path)
 				{
 					gwexe_modlist_original_contents = File.ReadAllText(gwexe_modlist_path);
 					File.WriteAllText(gwexe_modlist_path, "");
